+++ src/mappings/Common/Seasons/NameById.js
@@ -1,6 +1,7 @@
 const NameById = {
     '1': "2018",
-    '2': "2019"
+    '2': "2019",
+    '3': "2020",
 };
 
 module.exports = NameById;